+++ src/runtime/index.js
@@ -9,3 +9,4 @@ export { configure, __app_onerror, isFunction } from './utils';
 export * from '../parts/slot.runtime';
 export * from '../parts/keep-alive.runtime';
 export * from '../parts/select.runtime.js';
+export * from '../parts/radio.runtime.js';